Move util functions from excel reader to column parsing module
