+++ tables/readers/parsers/tests/test_column_parsers.py
@@ -6,8 +6,31 @@ from numpy.testing import assert_array_equal
 from pytest import raises
 
 from tables.readers.parsers import parse_column
-from tables.readers.parsers.columns import _parse_onoff_column, _parse_float_column, \
-    _parse_datetime_column, _parse_text_column
+from ..columns import normalize_if_str, is_missing_data_marker, _parse_onoff_column, \
+    _parse_float_column, _parse_datetime_column, _parse_text_column, parse_column
+
+
+def test_normalize_if_str():
+    assert normalize_if_str(" NoRmALiZe me\t") == "normalize me"
+    assert normalize_if_str("already normalized") == "already normalized"
+    assert normalize_if_str(42) == 42
+    assert normalize_if_str(None) is None
+
+
+@pytest.mark.parametrize(
+    "x,out",
+    [
+        ("-", True),
+        ("NaN", True),
+        ("nan", True),
+        ("NAN", True),
+        ("nAn", True),
+        ("Non!", False),
+        (None, False),
+    ],
+)
+def test_is_missing_data_marker(x, out):
+    assert is_missing_data_marker(x) == out
 
 
 def test__parse_text_column():
